+++ web-wallet/src/lib/stores/gasStore.js
@@ -1,8 +1,8 @@
 import { readable } from "svelte/store";
 
 const gasLimits = {
-  gasLimitLower: BigInt(import.meta.env.VITE_GAS_LIMIT_LOWER ?? 10000000),
-  gasLimitUpper: BigInt(import.meta.env.VITE_GAS_LIMIT_UPPER ?? 1000000000),
+  gasLimitLower: BigInt(import.meta.env.VITE_GAS_LIMIT_LOWER ?? 10_000_000),
+  gasLimitUpper: BigInt(import.meta.env.VITE_GAS_LIMIT_UPPER ?? 1_000_000_000),
   gasPriceLower: BigInt(import.meta.env.VITE_GAS_PRICE_LOWER ?? 1),
 };